修改设备使用记录使用时间改float

+++ tiiimmo/admin/src/main/java/com/linln/admin/device/VO/DeviceUseHistoryVO.java
@@ -25,7 +25,7 @@ public class DeviceUseHistoryVO {
     //使用人
     private String usePerson;
     //使用时间
-    private Integer useTime;
+    private Float useTime;
     //生产数量
     private Integer amount;
     //使用情况
